+++ ProjectManager/src/cmd/defines.c
@@ -14,24 +14,34 @@
 #   define OS_DELIMITER '/'
 #endif
 
-void cmd_define_add(cmd_args_t* argsptr, toml_section* cfg)
+cmd_args_t cmd_define_add(cmd_args_t* argsptr, toml_section* cfg)
 {
     cmd_args_t args = *argsptr;
     if (args.count < 1)
     {
         LOGERROR("Please specify define name and value (optional)");
-        return;
+        return EMPTY_CMD_ARGS;
     }
     const char* name = args.values[0];
     const char* value = args.values[1];
 
+    cmd_args_t endargs = {
+        args.count-1,
+        args.values+1,
+    };
+
     toml_property* define_property = toml_section_add_property(cfg, stringf("defines.%s", name));
     if (args.count > 1)
+    {
         toml_property_set_value_from_string(define_property, value);
+        endargs.count--;
+        endargs.values++;
+    }
 
     LOGINFO("Added define '%s'", name);
 
     save_project_file(cfg);
+    return endargs;
 }
 
 void cmd_define_remove(cmd_args_t* argsptr, toml_section* cfg)
@@ -56,37 +66,36 @@ void cmd_define_remove(cmd_args_t* argsptr, toml_section* cfg)
     save_project_file(cfg);
 }
 
-void cmd_defines(cmd_args_t* userparam)
+cmd_args_t cmd_defines(cmd_args_t* userparam)
 {
     cmd_args_t args = *userparam;
-    if (args.count == 0)
+    cmd_args_t endargs = args;
+    if (args.count <= 0)
         LOGINFO("Specify what to do, aka `add` or `remove`");
 
     toml_section cfg = TOML_SECTION_EMPTY;
     if (!open_project_file(&cfg))
     {
         LOGERROR("Cannot find project file in current directory");
-        return;
+        return EMPTY_CMD_ARGS;
     }
-    if (args.count == 0)
+    if (args.count <= 0)
         goto show_defines;
 
     toml_section* defines_section = NULL;
     const char* command = args.values[0];
-    cmd_args_t args2 = {args.count-1, args.values+1};
+    endargs = (cmd_args_t){args.count-1, args.values+1};
 
     if (strcmp(command, "add") == 0)
     {
-        cmd_define_add(&args2, &cfg);
+        endargs = cmd_define_add(&endargs, &cfg);
         toml_section_dispose(&cfg);
-        return;
+        return EMPTY_CMD_ARGS;
     }
     else if (strcmp(command, "remove") == 0)
-    {
-        cmd_define_remove(&args2, &cfg);
-        toml_section_dispose(&cfg);
-        return;
-    }
+        cmd_define_remove(&endargs, &cfg);
+    else
+        LOGERROR("Invalid command '%s'\n", command);
 
 show_defines:
     defines_section = toml_section_get_section(&cfg, "defines");
@@ -102,4 +111,5 @@ show_defines:
     }
 
     toml_section_dispose(&cfg);
+    return endargs;
 }